finish top and header
