--- src/modules/Home/components/description.js
@@ -1,9 +1,0 @@
-import React from 'react';
-
-const description = (props) => (
-  <div className="description">
-    <p>大学最後の１年間で学んだことをつらつらと書いていきたいブログ。</p>
-  </div>
-);
-
-export default description;
