+++ src/modules/Home/components/top/description.js
@@ -1,0 +1,9 @@
+import React from 'react';
+
+const description = (props) => (
+  <div id="description">
+    <p>大学最後の１年間で学んだことをつらつらと書いていきたいブログ</p>
+  </div>
+);
+
+export default description;
