+++ src/modules/Home/containers/homeContainer.js
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
-import Top from '../components/top.js'
+import Top from '../components/top/top.js'
+import Header from '../components/header/header.js'
 //import Profile from '../components/Profile.js'
 //import Article from '../components/Article.js'
 
@@ -7,8 +8,7 @@ export default class HomeContainers extends Component {
   render() {
     return (
       <div id="home">
-        {/*<Header />
-        */}
+        <Header />
         <Top />
         {/*<Profile />
           <Article />
